+++ CHANGELOG.md
@@ -8,6 +8,9 @@
 - Added variable extrusion width based on move type
 - Added Setting validator that check for common issues with settings
 - Added warnings
+- Added Acceleration, Feedrate, and Jerk settings
+- Added Retraction Wipe
+
 ### Fixes